First version of the programm
This first version contains the menu with its respective 6 options, where the most important is the sixth one: the Sort. Since I seek to test my knowledge on that subject, so that is the most important option, and the sort works with different types of sorting depending on the type of order you want, from using the name or the value of the object to sorting the data in ascending or descending order.